frontdesk_ai_supervisor/voice_agent/requirements_voice.txt
# LiveKit Voice Agent Dependencies
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
livekit-plugins-deepgram>=0.6.0
python-dotenv>=1.0.0
pyaudio>=0.2.14
